Kingdom command and AFK fixes

+++ src/main/java/com/roseworld/tab/Commands/KingdomCommand.java
@@ -13,7 +13,6 @@ import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.event.ClickEvent;
 import net.kyori.adventure.text.event.HoverEvent;
 import net.kyori.adventure.text.format.TextColor;
-import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 
 import java.util.List;
@@ -35,10 +34,19 @@ public class KingdomCommand {
                             String invite = StringArgumentType.getString(context, "kingdom");
                             if(KingdomHandler.getInvites().contains(invite)){
                                 KingdomHandler.acceptInvite(invite, player);
+                                return Command.SINGLE_SUCCESS;
                             }
+                            player.sendMessage(Message.Info("Invalid invite code!"));
                             return Command.SINGLE_SUCCESS;
                         }))
                 .then(Commands.literal("create")
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom == null;
+                            }
+                            return false;
+                        })
                         .then(Commands.argument("name", StringArgumentType.string())
                                 .executes(context -> {
                                     String name = StringArgumentType.getString(context, "name");
@@ -46,9 +54,10 @@ public class KingdomCommand {
                                         player.sendMessage(Message.Warning("Unable to create a kingdom cause the number of kingdoms exceeds the limit!"));
                                     }
                                     if(!KingdomHandler.isInKingdom(player)){
-                                        Kingdom kingdom = new Kingdom(name, player);
-                                        kingdom.createSeparator();
-                                        kingdom.joinKingdom(player);
+                                        Kingdom newKingdom = new Kingdom(name, player);
+                                        newKingdom.createSeparator();
+                                        newKingdom.joinKingdom(player);
+                                        player.updateCommands();
                                     }else{
                                         player.sendMessage(Message.Info("You are already in a kingdom!"));
                                         player.sendMessage(Message.Warning("You need to abandon your kingdom to create a new one."));
@@ -56,52 +65,65 @@ public class KingdomCommand {
                                     return Command.SINGLE_SUCCESS;
                                 })))
                 .then(Commands.literal("join")
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom == null;
+                            }
+                            return false;
+                        })
                         .then(Commands.argument("kingdom", StringArgumentType.string())
                                 .suggests((context, builder) -> {
-                                    KingdomHandler.getKingdoms().forEach(kingdom -> builder.suggest(kingdom.getName()));
+                                    KingdomHandler.getKingdoms().forEach(kingdom -> {
+                                        if(kingdom.isPublic()) builder.suggest(kingdom.getName());
+                                    });
                                     return builder.buildFuture();
                                 })
                                 .executes(context -> {
                                     String name = StringArgumentType.getString(context, "kingdom");
-                                    if (!KingdomHandler.isInKingdom(player)) {
-                                        for (Kingdom kingdom : KingdomHandler.getKingdoms()) {
-                                            if (kingdom.getName().equals(name))
-                                                if (kingdom.isPublic()) {
-                                                    kingdom.joinKingdom(player);
-                                                    return Command.SINGLE_SUCCESS;
-                                                } else {
-                                                    player.sendMessage(Message.Info("This kingdom is not public."));
-                                                    return Command.SINGLE_SUCCESS;
-                                                }
-                                        }
-                                        player.sendMessage(Message.Warning("There is not such kingdom!"));
-                                    }else{
-                                        player.sendMessage(Message.Info("You are already in a kingdom!"));
-                                        player.sendMessage(Message.Warning("You need to abandon your kingdom to join a new one."));
+                                    for (Kingdom kingdom : KingdomHandler.getKingdoms()) {
+                                        if (kingdom.getName().equals(name))
+                                            if (kingdom.isPublic()) {
+                                                kingdom.joinKingdom(player);
+                                                player.updateCommands();
+                                                return Command.SINGLE_SUCCESS;
+                                            } else {
+                                                player.sendMessage(Message.Info("This kingdom is not public."));
+                                                return Command.SINGLE_SUCCESS;
+                                            }
                                     }
+                                    player.sendMessage(Message.Warning("There is not such kingdom!"));
                                     return Command.SINGLE_SUCCESS;
                                 })))
                 .then(Commands.literal("leave")
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom != null;
+                            }
+                            return false;
+                        })
                         .executes(context -> {
                             Kingdom kingdom = KingdomHandler.getKingdom(player);
-                            if(kingdom == null) {
-                                player.sendMessage(Message.Info("You are not in a kingdom!"));
-                                return Command.SINGLE_SUCCESS;
-                            }
-                            player.sendMessage(Component.text("You left ").append(Component.text(kingdom.getName()).append(Component.text("!"))).color(TextColor.fromHexString("#5ae630")));
+                            if(kingdom == null) return Command.SINGLE_SUCCESS;
                             kingdom.leaveKingdom(player);
+                            player.updateCommands();
                             return Command.SINGLE_SUCCESS;
                         }))
                 .then(Commands.literal("delete")
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom != null && kingdom.getOwner().equals(player.getUniqueId());
+                            }
+                            return false;
+                        })
                         .executes(context -> {
                             Kingdom kingdom = KingdomHandler.getKingdom(player);
-                            if(kingdom == null) {
-                                player.sendMessage(Message.Info("You are not in a kingdom!"));
-                                return Command.SINGLE_SUCCESS;
-                            }
-                            if(kingdom.getOwner().equals(player.getUniqueId())) {
+                            if(kingdom != null && kingdom.getOwner().equals(player.getUniqueId())) {
                                 kingdom.deleteKingdom();
                                 player.sendMessage(Message.Warning(kingdom.getName() + " has been deleted!"));
+                                player.updateCommands();
                                 return Command.SINGLE_SUCCESS;
                             }else{
                                 player.sendMessage(Message.Warning("You don't have a permission to run this action!"));
@@ -109,52 +131,71 @@ public class KingdomCommand {
                             return Command.SINGLE_SUCCESS;
                         }))
                 .then(Commands.literal("public")
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom != null;
+                            }
+                            return false;
+                        })
                         .executes(context -> {
                             Kingdom kingdom = KingdomHandler.getKingdom(player);
-                            if(kingdom == null) {
-                                player.sendMessage(Message.Info("You are not in a kingdom!"));
-                                return Command.SINGLE_SUCCESS;
-                            }
-                            if(kingdom.isPublic()){
-                                player.sendMessage(Message.Info(kingdom.getName() + " is now public!"));
-                            }else {
-                                player.sendMessage(Message.Info(kingdom.getName() + " is now private!"));
+                            if(kingdom != null) {
+                                if(kingdom.isPublic()){
+                                    player.sendMessage(Message.Info(kingdom.getName() + " is public!"));
+                                }else {
+                                    player.sendMessage(Message.Info(kingdom.getName() + " is private!"));
+                                }
                             }
                             return Command.SINGLE_SUCCESS;
                         })
                         .then(Commands.argument("state", BoolArgumentType.bool())
+                                .requires(source -> {
+                                    if(source.getExecutor() instanceof Player king) {
+                                        Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                        return kingdom != null && kingdom.getOwner().equals(player.getUniqueId());
+                                    }
+                                    return false;
+                                })
                                 .executes(context -> {
                                     boolean state = BoolArgumentType.getBool(context, "state");
                                     Kingdom kingdom = KingdomHandler.getKingdom(player);
-                                    if(kingdom == null) {
-                                        player.sendMessage(Message.Info("You are not in a kingdom!"));
-                                        return Command.SINGLE_SUCCESS;
-                                    }
-                                    if(kingdom.getOwner().equals(player.getUniqueId())){
+                                    if(kingdom != null && kingdom.getOwner().equals(player.getUniqueId())){
                                         kingdom.setPublic(state);
-                                    }else{
-                                        player.sendMessage(Message.Warning("You don't have permissions to change the kingdom!"));
+                                        if(state){
+                                            player.sendMessage(Message.Info(kingdom.getName() + " is now public!"));
+                                        }else{
+                                            player.sendMessage(Message.Info(kingdom.getName() + " is now invite-only!"));
+                                        }
                                     }
                                     return Command.SINGLE_SUCCESS;
                                 })))
                 .then(Commands.literal("rename")
-                        .then(Commands.argument("name", StringArgumentType.string()))
-                        .executes(context -> {
-                            String newName = StringArgumentType.getString(context, "name");
-                            Kingdom kingdom = KingdomHandler.getKingdom(player);
-                            if(kingdom == null) {
-                                player.sendMessage(Message.Info("You are not in a kingdom!"));
-                                return Command.SINGLE_SUCCESS;
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom != null && kingdom.getOwner().equals(player.getUniqueId());
                             }
-                            if(kingdom.getOwner().equals(player.getUniqueId())){
-                                kingdom.setName(newName);
-                            }else{
-                                player.sendMessage(Message.Warning("You don't have permissions to change the kingdom!"));
-                            }
-                            return Command.SINGLE_SUCCESS;
-                        }))
+                            return false;
+                        })
+                        .then(Commands.argument("name", StringArgumentType.string())
+                                .executes(context -> {
+                                    String newName = StringArgumentType.getString(context, "name");
+                                    Kingdom kingdom = KingdomHandler.getKingdom(player);
+                                    if(kingdom != null && kingdom.getOwner().equals(player.getUniqueId())){
+                                        kingdom.setName(newName);
+                                    }
+                                    return Command.SINGLE_SUCCESS;
+                                })))
                 .then(Commands.literal("invite")
-                        .then(Commands.argument("player", ArgumentTypes.players())
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom != null;
+                            }
+                            return false;
+                        })
+                        .then(Commands.argument("player", ArgumentTypes.player())
                                 .executes(context -> {
                                     Kingdom kingdom = KingdomHandler.getKingdom(player);
                                     if(kingdom == null){
@@ -175,6 +216,13 @@ public class KingdomCommand {
                                     return Command.SINGLE_SUCCESS;
                                 })))
                 .then(Commands.literal("chat")
+                        .requires(source -> {
+                            if(source.getExecutor() instanceof Player king) {
+                                Kingdom kingdom = KingdomHandler.getKingdom(king);
+                                return kingdom != null;
+                            }
+                            return false;
+                        })
                         .executes(context -> {
                             Kingdom kingdom = KingdomHandler.getKingdom(player);
                             if(kingdom != null){
@@ -187,8 +235,6 @@ public class KingdomCommand {
                                 kingdom.leaveChat(player);
                                 KingdomHandler.removeKingdomChatter(player);
                                 player.sendMessage(Component.text("Switched to the general chat!", TextColor.fromHexString("#5ae630")));
-                            }else {
-                                player.sendMessage(Message.Warning("You're not in a kingdom!"));
                             }
                             return Command.SINGLE_SUCCESS;
                         }))
